New common library witch will be used by digiKam and kipi-plugins to handle Pictures metadata (Exif/Iptc).
This shared library is a wrapper around Exiv2 library. It give a Qt like interface to control
Exiv2 actions on pictures. Using this library will remove all duplicate code on kipi-plugins::exiv2iface and
digikam::dmetadata.

The current implementation is the same than kipi-plugins/exiv2iface. This library is not yet used in digiKam/kipi-plugins.
We will added a new depency to this library in kipi-plugins in first and later to digiKam.

CCMAIL: digikam-devel@kde.org, kde-imaging@kde.org, marcel.wiesweg@gmx.de


svn path=/trunk/extragear/libs/libkexiv2/; revision=627390