Improve merging of sidecar with image metadata.
If there is a sidecar, image metadata may remain, but can be outdated if image shall not
or cannot be edited. Even more, a field can be removed from a sidecar as part of an explicit
editing operation, in which case it may remain in the image metadata, but needs to be ignored.

For those fields were a mapping between EXIF, IPTC to XMP is cleanly defined,
implement merging which takes into account the points listed above.

BUG: 309058